run RequireAWSCredentials if using interactive CLI flow to setup provider
